requirements.txt
# Core LangChain ecosystem (LCEL + agents)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Google Gemini / Google AI
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

# Tavily search tool
tavily-python>=0.3.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Async support (important for agents/tools)
httpx>=0.27.0
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Data handling (optional but useful for parsing outputs)
pydantic>=2.7.0

# Logging / debugging
rich>=13.7.0

# Token counting (helps with LLM usage control)
tiktoken>=0.7.0

# Optional: faster JSON handling
orjson>=3.10.0

# Optional: retry logic for APIs
tenacity>=8.2.3